feat(payments): extend Money VO with currency-aware precision and move CurrencyVO to shared
- Move CurrencyVO from events/domain to @shared/domain (multi-module usage)
- Events module re-exports from @shared for backward compatibility
- Money VO now rounds based on currency decimals (TND=3, USD/EUR=2)
- Add toSmallestUnit(), toCents(), toMillimes(), fromSmallestUnit()
- Add percentage() method for commission calculation
- Add isGreaterThanOrEqual(), isLessThanOrEqual() comparisons
- Validate currency before rounding (prevents crash on invalid input)
- Add CurrencyVO.toSmallestUnit() and fromSmallestUnit() static helpers
- Update test to reflect correct TND 3-decimal precision
- All 1484 existing tests pass (zero regressions)

Issue: Payments Module - Issue #1
