+++ backend/src/modules/events/domain/value-objects/currency.vo.ts
@@ -1,245 +1,14 @@
 /**
- * Currency Enum
+ * Currency VO - Re-exported from @shared/domain/value-objects/currency.vo
  *
- * Defines the supported currencies in the system.
- * Primary currency is TND (Tunisian Dinar), with support for EUR and USD.
+ * This file exists for backward compatibility. The canonical source
+ * is now in @shared since multiple modules (Events, Tickets, Payments) use it.
  */
-export enum Currency {
-  TND = 'TND',
-  EUR = 'EUR',
-  USD = 'USD',
-}
+export {
+  Currency,
+  CurrencyVO,
+  CURRENCY_METADATA,
+} from '@shared/domain/value-objects/currency.vo';
 
-/**
- * Currency metadata for display and formatting (not exported to comply with architecture)
- */
-interface CurrencyMetadata {
-  code: Currency;
-  symbol: string;
-  name: string;
-  nameFr: string;
-  decimals: number;
-  locale: string;
-}
+export type { CurrencyMetadata } from '@shared/domain/value-objects/currency.vo';
 
-/**
- * Currency metadata mapping
- */
-export const CURRENCY_METADATA: Record<Currency, CurrencyMetadata> = {
-  [Currency.TND]: {
-    code: Currency.TND,
-    symbol: 'DT',
-    name: 'Tunisian Dinar',
-    nameFr: 'Dinar Tunisien',
-    decimals: 3, // TND has 3 decimal places
-    locale: 'ar-TN',
-  },
-  [Currency.EUR]: {
-    code: Currency.EUR,
-    symbol: '€',
-    name: 'Euro',
-    nameFr: 'Euro',
-    decimals: 2,
-    locale: 'fr-FR',
-  },
-  [Currency.USD]: {
-    code: Currency.USD,
-    symbol: '$',
-    name: 'US Dollar',
-    nameFr: 'Dollar américain',
-    decimals: 2,
-    locale: 'en-US',
-  },
-};
-
-/**
- * Currency Value Object
- *
- * Provides currency validation and formatting utilities
- */
-export class CurrencyVO {
-  private constructor(private readonly currency: Currency) {}
-
-  /**
-   * Get the currency value
-   */
-  get value(): Currency {
-    return this.currency;
-  }
-
-  /**
-   * Create a CurrencyVO from string
-   */
-  static create(currency: string): CurrencyVO {
-    const normalized = currency.toUpperCase().trim();
-    if (!Object.values(Currency).includes(normalized as Currency)) {
-      throw new Error(
-        `Invalid currency: ${currency}. Must be one of: ${Object.values(Currency).join(', ')}`,
-      );
-    }
-    return new CurrencyVO(normalized as Currency);
-  }
-
-  /**
-   * Create from enum directly
-   */
-  static fromEnum(currency: Currency): CurrencyVO {
-    return new CurrencyVO(currency);
-  }
-
-  /**
-   * Get the default currency (TND)
-   */
-  static default(): CurrencyVO {
-    return new CurrencyVO(Currency.TND);
-  }
-
-  /**
-   * Get currency metadata
-   */
-  getMetadata(): CurrencyMetadata {
-    return CURRENCY_METADATA[this.currency];
-  }
-
-  /**
-   * Get currency symbol
-   */
-  getSymbol(): string {
-    return CURRENCY_METADATA[this.currency].symbol;
-  }
-
-  /**
-   * Get currency name
-   */
-  getName(locale: 'en' | 'fr' = 'en'): string {
-    const metadata = CURRENCY_METADATA[this.currency];
-    return locale === 'fr' ? metadata.nameFr : metadata.name;
-  }
-
-  /**
-   * Get number of decimal places
-   */
-  getDecimals(): number {
-    return CURRENCY_METADATA[this.currency].decimals;
-  }
-
-  /**
-   * Format amount with this currency
-   */
-  format(amount: number): string {
-    const metadata = CURRENCY_METADATA[this.currency];
-    return new Intl.NumberFormat(metadata.locale, {
-      style: 'currency',
-      currency: this.currency,
-      minimumFractionDigits: metadata.decimals,
-      maximumFractionDigits: metadata.decimals,
-    }).format(amount);
-  }
-
-  /**
-   * Round amount to currency's decimal places
-   */
-  roundAmount(amount: number): number {
-    const decimals = CURRENCY_METADATA[this.currency].decimals;
-    const factor = Math.pow(10, decimals);
-    return Math.round(amount * factor) / factor;
-  }
-
-  /**
-   * Check if TND
-   */
-  isTND(): boolean {
-    return this.currency === Currency.TND;
-  }
-
-  /**
-   * Check if EUR
-   */
-  isEUR(): boolean {
-    return this.currency === Currency.EUR;
-  }
-
-  /**
-   * Check if USD
-   */
-  isUSD(): boolean {
-    return this.currency === Currency.USD;
-  }
-
-  /**
-   * Compare with another currency
-   */
-  equals(other: CurrencyVO): boolean {
-    return this.currency === other.currency;
-  }
-
-  /**
-   * String representation
-   */
-  toString(): string {
-    return this.currency;
-  }
-
-  // Static utility methods
-
-  /**
-   * Get all supported currencies
-   */
-  static getAllCurrencies(): Currency[] {
-    return Object.values(Currency);
-  }
-
-  /**
-   * Check if a value is a valid currency
-   */
-  static isValidCurrency(value: string): value is Currency {
-    return Object.values(Currency).includes(value.toUpperCase() as Currency);
-  }
-
-  /**
-   * Get symbol for a currency enum value
-   */
-  static getSymbolFor(currency: Currency): string {
-    return CURRENCY_METADATA[currency].symbol;
-  }
-
-  /**
-   * Format amount with currency enum value
-   */
-  static formatAmount(amount: number, currency: Currency): string {
-    const metadata = CURRENCY_METADATA[currency];
-    return new Intl.NumberFormat(metadata.locale, {
-      style: 'currency',
-      currency: currency,
-      minimumFractionDigits: metadata.decimals,
-      maximumFractionDigits: metadata.decimals,
-    }).format(amount);
-  }
-
-  /**
-   * Round amount to currency's decimal places
-   */
-  static roundAmountFor(amount: number, currency: Currency): number {
-    const decimals = CURRENCY_METADATA[currency].decimals;
-    const factor = Math.pow(10, decimals);
-    return Math.round(amount * factor) / factor;
-  }
-
-  /**
-   * Parse currency from string (case-insensitive), returns null if invalid
-   */
-  static fromString(value: string): CurrencyVO | null {
-    try {
-      return CurrencyVO.create(value);
-    } catch {
-      return null;
-    }
-  }
-
-  /**
-   * Get the default currency enum value
-   */
-  static getDefault(): Currency {
-    return Currency.TND;
-  }
-}
