+++ backend/src/shared/domain/index.ts
@@ -9,6 +9,7 @@ export * from './result';
 export * from './value-objects/email.vo';
 export * from './value-objects/phone.vo';
 export * from './value-objects/money.vo';
+export * from './value-objects/currency.vo';
 export * from './value-objects/date-range.vo';
 export * from './value-objects/url.vo';
