+++ backend/src/shared/domain/value-objects/money.vo.ts
@@ -1,5 +1,7 @@
 import { ValueObject } from '../value-object.base';
 
+import { CURRENCY_METADATA, CurrencyVO } from './currency.vo';
+
 interface MoneyProps {
   amount: number;
   currency: string;
@@ -14,12 +16,13 @@ export class InvalidMoneyException extends Error {
 
 /**
  * Money Value Object
- * 
- * Handles monetary values with currency
+ *
+ * Handles monetary values with currency-aware precision.
+ * TND uses 3 decimal places (millimes), USD/EUR use 2 (cents).
+ * Precision is driven by CURRENCY_METADATA — adding a new currency
+ * requires zero changes to this class.
  */
 export class Money extends ValueObject<MoneyProps> {
-  private static readonly SUPPORTED_CURRENCIES = ['TND', 'EUR', 'USD'];
-
   get amount(): number {
     return this.props.amount;
   }
@@ -29,17 +32,28 @@ export class Money extends ValueObject<MoneyProps> {
   }
 
   get formatted(): string {
-    const formatter = new Intl.NumberFormat('fr-TN', {
-      style: 'currency',
-      currency: this.props.currency,
-    });
-    return formatter.format(this.props.amount);
+    return CurrencyVO.formatAmount(this.amount, this.currency as any);
+  }
+
+  /**
+   * Get the number of decimal places for this money's currency
+   */
+  get decimals(): number {
+    const meta = CURRENCY_METADATA[this.currency as keyof typeof CURRENCY_METADATA];
+    return meta ? meta.decimals : 2;
   }
 
   static create(amount: number, currency: string = 'TND'): Money {
+    const normalized = currency.toUpperCase();
+    if (!CurrencyVO.isValidCurrency(normalized)) {
+      throw new InvalidMoneyException(
+        `Unsupported currency: ${normalized}. Supported: ${CurrencyVO.getAllCurrencies().join(', ')}`,
+      );
+    }
+    const rounded = CurrencyVO.roundAmountFor(amount, normalized as any);
     return new Money({
-      amount: Math.round(amount * 100) / 100, // Round to 2 decimals
-      currency: currency.toUpperCase(),
+      amount: rounded,
+      currency: normalized,
     });
   }
 
@@ -59,6 +73,14 @@ export class Money extends ValueObject<MoneyProps> {
 
   multiply(factor: number): Money {
     return Money.create(this.amount * factor, this.currency);
+  }
+
+  /**
+   * Calculate a percentage of this money amount
+   * @param rate - decimal rate (e.g., 0.04 for 4%)
+   */
+  percentage(rate: number): Money {
+    return Money.create(this.amount * rate, this.currency);
   }
 
   isPositive(): boolean {
@@ -83,6 +105,50 @@ export class Money extends ValueObject<MoneyProps> {
     return this.amount < other.amount;
   }
 
+  isGreaterThanOrEqual(other: Money): boolean {
+    this.assertSameCurrency(other);
+    return this.amount >= other.amount;
+  }
+
+  isLessThanOrEqual(other: Money): boolean {
+    this.assertSameCurrency(other);
+    return this.amount <= other.amount;
+  }
+
+  /**
+   * Convert to the smallest unit for this currency.
+   * TND → millimes (×1000), USD/EUR → cents (×100).
+   * Used when sending amounts to payment gateways.
+   */
+  toSmallestUnit(): number {
+    return CurrencyVO.toSmallestUnit(this.amount, this.currency as any);
+  }
+
+  /**
+   * Convenience alias for Stripe API (expects cents for 2-decimal currencies).
+   * For TND (3 decimals), still returns millimes (×1000).
+   */
+  toCents(): number {
+    return this.toSmallestUnit();
+  }
+
+  /**
+   * Convenience alias for Konnect API (expects millimes for TND).
+   * Equivalent to toSmallestUnit() for TND.
+   */
+  toMillimes(): number {
+    return this.toSmallestUnit();
+  }
+
+  /**
+   * Create Money from smallest unit (cents/millimes) back to decimal.
+   */
+  static fromSmallestUnit(smallestUnit: number, currency: string = 'TND'): Money {
+    const normalized = currency.toUpperCase();
+    const amount = CurrencyVO.fromSmallestUnit(smallestUnit, normalized as any);
+    return Money.create(amount, normalized);
+  }
+
   private assertSameCurrency(other: Money): void {
     if (this.currency !== other.currency) {
       throw new InvalidMoneyException(
@@ -95,9 +161,9 @@ export class Money extends ValueObject<MoneyProps> {
     if (typeof props.amount !== 'number' || isNaN(props.amount)) {
       throw new InvalidMoneyException('Amount must be a valid number');
     }
-    if (!Money.SUPPORTED_CURRENCIES.includes(props.currency)) {
+    if (!CurrencyVO.isValidCurrency(props.currency)) {
       throw new InvalidMoneyException(
-        `Unsupported currency: ${props.currency}. Supported: ${Money.SUPPORTED_CURRENCIES.join(', ')}`,
+        `Unsupported currency: ${props.currency}. Supported: ${CurrencyVO.getAllCurrencies().join(', ')}`,
       );
     }
   }
